+++ eventstore-mongodb/src/main/java/com/marketplace/eventstore/mongodb/MongoEventEntity.java
@@ -17,7 +17,7 @@ public interface MongoEventEntity {
 
   List<TypedEvent> getEvents();
 
-  int getVersion();
+  long getVersion();
 
   Instant getCreatedAt();
 }
